feat: Add start screen. Change the Player instantiation.
I made the startScreen function and implement it's invokation
at the init() of the Engine.js. I did it because to my opinion there is
no need to use any certain class to do such small piece of work here -
at the begining. Further in the game this function will not be used
again.

The Player instantiation is transfered to the startScreen().

Any relative small fixes across the code in the app.js.